cleaning up dependencies for spec conversion functions
