+++ packages/scripts/src/convert.ts
@@ -24,57 +24,10 @@
  * THE SOFTWARE.
  */
 
-import { ISDKConfigProps } from './sdkConfig'
-import { openApiFileName, logFetch } from './fetchSpec'
 import { log, success, warn } from '@looker/sdk-codegen-utils'
 import { fail, isFileSync, quit, readFileSync, run, utf8Encoding } from './nodeUtils'
 import * as fs from 'fs'
 import { ParameterStyle } from 'openapi3-ts'
-
-/*
-const { Spectral } = require('@stoplight/spectral')
-const { getLocationForJsonPath, parseWithPointers } = require('@stoplight/json')
-
-const lintyFresh = true
-
-const lintCheck = async (fileName: string) => {
-  if (!lintyFresh) return ''
-  // return `${fileName} lint checking was skipped`
-  try {
-    // const linter = run('speccy', ['lint', fileName])
-    const linter = new Spectral()
-    if (!linter) return fail('Lint', 'no response')
-    const spec = parseWithPointers(readFileSync(fileName))
-    linter
-      .run({
-        parsed: spec,
-        getLocationForJsonPath,
-      })
-      .then(console.log)
-    return ''
-    // if (
-    //   linter.toString().indexOf('Specification is valid, with 0 lint errors') >=
-    //   0
-    // ) {
-    //   return
-    // }
-  } catch (e) {
-    return quit(e)
-  }
-}
-
-*/
-
-/**
- * Checks OpenAPI file for lint errors
- *
- * NOTE: Currently disabled
- * @param {string} fileName
- * @returns {Promise<void>}
- */
-const lintCheck = async (fileName: string) => {
-  return ''
-}
 
 /**
  * Replaces x-looker-nullable with nullable for parameters and properties in a string
@@ -241,24 +194,4 @@ export const convertSpec = (swaggerFilename: string, openApiFilename: string) =>
   } catch (e) {
     return quit(e)
   }
-}
-
-/**
- * Fetch (if needed) and convert a Swagger API specification to OpenAPI
- * @param {string} name base name of the target file
- * @param {ISDKConfigProps} props SDK configuration properties to use
- * @returns {Promise<string>} name of converted OpenAPI file
- */
-export const logConvert = async (name: string, props: ISDKConfigProps) => {
-  const oaFile = openApiFileName(name, props)
-  if (isFileSync(oaFile)) return oaFile
-
-  const specFile = await logFetch(name, props)
-  const openApiFile = convertSpec(specFile, oaFile)
-  if (!openApiFile) {
-    return fail('logConvert', 'No file name returned for openAPI upgrade')
-  }
-
-  await lintCheck(openApiFile)
-  return openApiFile
 }
