+++ packages/scripts/src/sdkGen.ts
@@ -27,10 +27,9 @@
 import * as fs from 'fs'
 import { ISDKConfigProps, SDKConfig } from './sdkConfig'
 import { danger, log } from '@looker/sdk-codegen-utils'
-import { fetchLookerVersion, openApiFileName } from './fetchSpec'
+import { fetchLookerVersion, logConvert, openApiFileName } from './fetchSpec'
 import { MethodGenerator, specFromFile, StreamGenerator, TypeGenerator } from './sdkGenerator'
 import { getFormatter, Languages } from './languages'
-import { logConvert } from './convert'
 import { IVersionInfo, ICodeGen } from '@looker/sdk-codegen'
 import { FilesFormatter } from './reformatter'
 import { isDirSync, quit } from './nodeUtils'
